feat(pie-icons): DSW-000 refactor icons-update workflow into a script
