--- .github/workflows/update-icons.yml
@@ -1,46 +1,0 @@
-name: Update icons
-on:
-  workflow_dispatch:
-
-env:
-  HUSKY: 0
-
-jobs:
-  update-icons:
-    runs-on: ubuntu-latest
-    permissions:
-      contents: write
-      pull-requests: write
-    steps:
-      # Checkout the Repo
-      - name: Checkout
-        uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
-
-      - name: Setup Repo
-        uses: ./.github/actions/setup-repo
-        with:
-          node-version: 20
-          os: ubuntu-latest
-
-      - name: Install dependencies
-        shell: bash
-        run: yarn
-
-      - name: Update icons
-        run: yarn workspace @justeattakeaway/pie-icons update-icons
-
-      - name: Create PR
-        uses: actions/github-script@60a0d83039c74a4aee543508d2ffcb1c3799cdea # v7.0.1
-        env:
-          BRANCH_NAME: ${{ env.BRANCH_NAME }} # this is set by the previous step
-          CHANGESET_FILE_PATH: ${{ env.CHANGESET_FILE_PATH }} # this is set by the previous step
-        with:
-          script: |
-            const script = require('./packages/tools/pie-monorepo-utils/create-icons-update-pr/create-icons-update-pr.js');
-            const { BRANCH_NAME, CHANGESET_FILE_PATH } = process.env
-            const config = {
-              github,
-              branchName: BRANCH_NAME,
-              changesetFilePath:CHANGESET_FILE_PATH
-            };
-            await script(config);
