+++ packages/tools/pie-icons/package.json
@@ -38,6 +38,7 @@
     "glob": "11.0.0",
     "is-svg": "5.1.0",
     "lint-staged": "12.5.0",
+    "open": "10.1.0",
     "slugify": "1.6.6",
     "svgo": "3.0.2",
     "webpack": "5.94.0",
